158-ASS2/requirements.txt
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.30.0
pillow>=10.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jupyter>=1.0.0
ipykernel>=6.25.0
tqdm>=4.66.0

